+++ lib/screens/screens.dart
@@ -4,3 +4,4 @@ export 'auth_page.dart';
 export 'bottom_tab_page.dart';
 export 'profile_page.dart';
 export 'profile_modify_page.dart';
+export 'fav_page.dart';
