+++ lib/widgets/widgets.dart
@@ -1,5 +1,3 @@
 export 'auth/login.dart';
 export 'auth/register.dart';
-export 'modal_bottom_sheet.dart';
 export 'connection_widget.dart';
-export 'charger_widget.dart';
